# Much faster prompt processing for I-quants (ARM_NEON)